rename additional to contrib and make it visible
