+++ crates/jpreprocess-njd/src/lib.rs
@@ -1,4 +1,4 @@
-mod additional;
+mod contrib;
 mod node;
 mod open_jtalk;
 
@@ -7,6 +7,7 @@ use jpreprocess_dictionary::DictionaryFetcher;
 use jpreprocess_window::{IterQuintMut, IterQuintMutTrait};
 use lindera_tokenizer::token::Token;
 
+pub use contrib::*;
 pub use node::*;
 pub use open_jtalk::*;
